+++ tests/test_html.py
@@ -276,11 +276,11 @@ def test_markdown(app):
     print(view.render(source="\n- Item 1\n- Item 2\n"))
     assert len(view.proxy.widget.xpath("/html/body/div/ul/li")) == 2
 
-    #for e in evts:
-        #print(e)
-    #e = evts[-1]["value"]
-    #assert e["name"] == "source"
-    #assert e["value"] == view.md.render()
+    for e in evts:
+        print(e)
+    e = evts[-1]["value"]
+    assert e["name"] == "source"
+    assert e["value"] == view.md.render()
 
 
 def test_markdown_proxy(app):